fix runaway task
